requirements.txt
feedparser==6.0.11
flask-cors==6.0.1
flask==3.0.3
langchain-core==0.3.74
langchain-openai==0.3.30
numpy==1.26.4
openai-whisper==20250625
pinecone-client==7.3.0
pydantic==2.11.7
pydub==0.25.1
python-dotenv==0.21.0
requests==2.32.4
sentence-transformers==5.1.0
torch==2.7.1
